+++ Tokens/ButterToken.sol
@@ -93,15 +93,15 @@ contract ButterToken { //ButterToken == GTK
     }
 
     // VIEWERS
-    function getInitialSupply() external view returns (uint256) {
+    function getInitialSupply() external view returns (uint) {
         return initialTokens;
     }
 
-    function getTotalSupply() external view returns (uint256) {
+    function getTotalSupply() external view returns (uint) {
         return totalTokens;
     }
 
-    function getBalanceOf() external view returns (uint256) {
+    function getBalanceOf() external view returns (uint) {
         return balances[msg.sender];
     }
